separate compile vs consolidate
